server/requirements.txt
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.1.2
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1